backend/python/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cryptohub-backend"
version = "0.1.0"
description = "CryptoHub Python Backend - AI & Quantitative Trading Engine"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "celery[redis]>=5.4.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "ccxt>=4.4.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "openai>=1.60.0",
    "scikit-learn>=1.6.0",
    "sqlalchemy>=2.0.36",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "grpcio>=1.69.0",
    "grpcio-tools>=1.69.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "asyncpg>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]